Added new option
NET_OPT_EXECUTE_PACKET_ASYNC

this option will spawn a new thread to execute the packet in.
this option will allow the receive thread to keep executing income packets, even if a packet callback is currently on hold due to a blocked action.

fixed bug in Net::Json where a shared memory object/array was destroyed even when some other object was holding the ownership of that pointer.